Remove temporary files after each plot.
If the user needs to use a persistent temporary file, it can be created
manually. This should not normally be necessary, as the Plot class now
offers easy replotting of data (no need to rely on Gnuplot's `replot');
if one wants to reuse a file as a performance hack, they are free to
hack.